Blog system in progress - partial refactor
